Задание №5. Переписывание классов с учетом наследования от абстрактного компанента
