+++ src/components/film-detail.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import AbstractComponent from "./abstract-component.js";
 
 const createComment = (comment) => {
   const {text, autor, date, emoji, img} = comment;
@@ -145,25 +145,13 @@ const createFilmDetailTemplate = (film) => {
   );
 };
 
-export default class FilmDetail {
+export default class FilmDetail extends AbstractComponent {
   constructor(film) {
+    super();
     this._film = film;
-    this._element = null;
   }
 
   getTemplate() {
     return createFilmDetailTemplate(this._film);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
