+++ src/components/films-fail-section.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import AbstractComponent from "./abstract-component.js";
 
 const createFailFilmSectionTemplate = () => {
   return (
@@ -8,24 +8,8 @@ const createFailFilmSectionTemplate = () => {
   );
 };
 
-export default class FailFilmSection {
-  constructor() {
-    this._element = null;
-  }
-
+export default class FailFilmSection extends AbstractComponent {
   getTemplate() {
     return createFailFilmSectionTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
