+++ src/components/films-list-section.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import AbstractComponent from "./abstract-component.js";
 
 const createFilmsListSectionTemplate = () => {
   return (
@@ -9,24 +9,8 @@ const createFilmsListSectionTemplate = () => {
   );
 };
 
-export default class FailFilmSection {
-  constructor() {
-    this._element = null;
-  }
-
+export default class FailFilmSection extends AbstractComponent {
   getTemplate() {
     return createFilmsListSectionTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
