+++ src/components/films-section.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import AbstractComponent from "./abstract-component.js";
 
 const createFilmsSectionTemplate = () => {
   return (
@@ -6,24 +6,8 @@ const createFilmsSectionTemplate = () => {
   );
 };
 
-export default class FilmSection {
-  constructor() {
-    this._element = null;
-  }
-
+export default class FilmSection extends AbstractComponent {
   getTemplate() {
     return createFilmsSectionTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
