+++ src/components/show-more-button.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import AbstractComponent from "./abstract-component.js";
 
 const createShomMoreButton = () => {
   return (
@@ -6,24 +6,8 @@ const createShomMoreButton = () => {
   );
 };
 
-export default class ShomMoreButton {
-  constructor() {
-    this._element = null;
-  }
-
+export default class ShomMoreButton extends AbstractComponent {
   getTemplate() {
     return createShomMoreButton();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
