+++ src/utils.js
@@ -38,32 +38,4 @@ const getRandomMixedArray = (array, number) => {
   return mixedArray;
 };
 
-//  --------------------  Функция по созданию DOM элемента
-
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-//  --------------------  Функция добавления DOM элемента в разметку
-
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export {getRandomIntegerNumber, getRandomArrayItem, getRandomNumbersArray, getRandomMixedArray, createElement, RenderPosition, render};
-
+export {getRandomIntegerNumber, getRandomArrayItem, getRandomNumbersArray, getRandomMixedArray};
